eckit::maths::ConvexHull move exceptions from maths::Qhull to maths::ConvexHull to be abstract from dependency
